+++ xenian_channel/bot/models/user_state.py
@@ -1,6 +1,6 @@
 from threading import Lock
 
-from mongoengine import Document, NULLIFY, ReferenceField, StringField
+from mongoengine import Document, NULLIFY, ReferenceField, StringField, DictField
 
 from xenian_channel.bot.models import ChannelSettings
 from xenian_channel.bot.models.tg_user import TgUser
@@ -18,12 +18,17 @@ class UserState(Document):
     CHANGE_DEFAULT_CAPTION = 'change default caption'
     CHANGE_DEFAULT_REACTION = 'change default reaction'
     CREATE_SINGLE_POST = 'create single post'
+    SCHEDULE_ADDED_MESSAGES_WHEN = 'schedule added messages when'
+    SCHEDULE_ADDED_MESSAGES_DELAY = 'schedule added messages delay'
+    SCHEDULE_ADDED_MESSAGES_BATCH = 'schedule added messages batch'
+    SCHEDULE_ADDED_MESSAGES_CONFIRMATION = 'schedule added messages confirmation'
     SEND_LOCKED = 'send_locked'
 
     user = ReferenceField(TgUser)
     state = StringField(default=IDLE)
 
     current_channel = ReferenceField(ChannelSettings, reverse_delete_rule=NULLIFY)
+    state_data = DictField(default={})
 
     save_lock = Lock()